AmplifyBase in lib-builtin-effects
