+++ libraries/CMakeLists.txt
@@ -73,6 +73,7 @@ set( LIBRARIES
    lib-preference-pages
    lib-dynamic-range-processor
    lib-audacity-application-logic
+   lib-builtin-effects
 )
 
 if ( ${_OPT}use_lv2 )
